adapted new api for export keys, removed keysets export keys menu and screens
