+++ android/src/main/java/io/parity/signer/screens/keysets/KeySetsNavSubgraph.kt
@@ -3,7 +3,6 @@ package io.parity.signer.screens.keysets
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.State
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
@@ -12,9 +11,6 @@ import androidx.navigation.compose.rememberNavController
 import io.parity.signer.domain.KeySetsSelectModel
 import io.parity.signer.domain.Navigator
 import io.parity.signer.domain.NetworkState
-import io.parity.signer.screens.keysets.export.KeySetsExportScreenFull
-import io.parity.signer.ui.BottomSheetWrapperRoot
-import io.parity.signer.uniffi.Action
 
 /**
  * Navigation Subgraph with compose nav controller for those Key Set screens which are not part of general
@@ -42,39 +38,10 @@ fun KeySetsNavSubgraph(
 				)
 			}
 		}
-		composable(KeySetsNavSubgraph.homeMenu) {
-			Box(modifier = Modifier.statusBarsPadding()) {
-				KeySetsScreen(
-					model = model,
-					rootNavigator = rootNavigator,
-					localNavigator = navController,
-					networkState = networkState,
-				)
-			}
-			if (model.keys.isNotEmpty()) {
-				BottomSheetWrapperRoot(onClosedAction = {
-					navController.popBackStack(KeySetsNavSubgraph.home, false)
-				}) {
-					KeySetsMenuBottomSheet(navigator = navController)
-				}
-			} else {
-				navController.popBackStack(KeySetsNavSubgraph.home, false)
-			}
-		}
-		composable(KeySetsNavSubgraph.export) {
-			KeySetsExportScreenFull(
-				model = model,
-				onClose = {
-					navController.popBackStack(KeySetsNavSubgraph.home, false)
-				},
-			)
-		}
 	}
 }
 
 
 internal object KeySetsNavSubgraph {
 	const val home = "keysets_home"
-	const val homeMenu = "keysets_menu"
-	const val export = "keysets_export"
 }
